+++ coding-backend/src/main/java/com/codingbackend/domain/cart/CartMapper.java
@@ -17,13 +17,6 @@ public interface CartMapper {
     @Select("SELECT COUNT(*) FROM cart WHERE restaurant_id=#{restaurantId} AND menu_name = #{menuName} And user_id=#{userId} AND payment_status = FALSE")
     int selectByRestaurantIdAndMenuNameAndUserIdAndPaymentStatusFalse(Long restaurantId, String menuName, Integer userId);
 
-    @Update("""
-            UPDATE cart
-            SET menu_count = #{menuCount}, total_price = #{totalPrice}
-            WHERE restaurant_id=#{restaurantId} AND menu_name = #{menuName}
-            """)
-    void update(Cart cart);
-
     @Delete("DELETE FROM cart WHERE restaurant_id=#{restaurantId} AND user_id=#{userId}")
     int deleteByRestaurantIdAndUserId(Long restaurantId, Integer userId);
 
@@ -61,4 +54,14 @@ public interface CartMapper {
               AND payment_status = FALSE
             """)
     List<Cart> selectByUserIdAndRestaurantIdAndPaymentStatus(Integer userId, Long restaurantId);
+
+    @Select("""
+            SELECT id, restaurant_id, user_id, 
+                   menu_name, menu_count, menu_price, 
+                   total_price, inserted, payment_status 
+            FROM cart 
+            WHERE user_id=#{userId}
+              AND payment_status = TRUE
+            """)
+    List<Cart> selectByUserIdAndPaymentStatusTrue(Integer userId);
 }
